Ephemeral Chain Runners (#587)
* Init cltester setup

* Setup Chain build step

* Setup genesis ephemeral data

* Complete genesis runner

* Adjust build step name

* Complete nodeos runner

* Fix runners test cases

* fix  atomic market files

* Fix chain runners
